Add Query and View interfaces
