+++ server/src/main/java/io/javaclasses/filehub/api/Query.java
@@ -1,0 +1,7 @@
+package io.javaclasses.filehub.api;
+
+/**
+ * An abstract base which represents intention of a client to retrieve any data.
+ */
+public interface Query {
+}
